requirements.txt
livekit >= 0.18.2
livekit-agents>=0.12.1
livekit-plugins-openai>=0.10.9
livekit-plugins-deepgram>=0.6.13
livekit-plugins-silero>=0.7.4
livekit-plugins-turn-detector>=0.3.1
python-dotenv~=1.0
